# perf: Make validation check in regressor_column_matrix 300x faster